Create fiveNameGenerators
This code has the building blocks for five "types" of name generators. Using this as a starting point, you should be able to make much more complex names, such as those involving more syllables, names with clan names or last names, and names with A and A' components. 

This code is NOT ready to run out of the box. You MUST add the types of syllables, vowels, consonants, etc. that you want to see in your names. This code is ONLY how to combine your chosen syllables and letters to create names.